Add comprehensive Services, Controllers, and DTOs for Customers, Suppliers, Staff and additional tables

+++ PowerTraderPOS.API/Services/Interfaces/ICustomerService.cs
@@ -5,10 +5,10 @@ namespace PowerTraderPOS.API.Services.Interfaces
     public interface ICustomerService
     {
         Task<IEnumerable<CustomerDto>> GetAllCustomersAsync();
-        Task<CustomerDto?> GetCustomerByIdAsync(int id);
+        Task<CustomerDto?> GetCustomerByIdAsync(decimal id);
         Task<CustomerDto> CreateCustomerAsync(CreateCustomerDto dto);
-        Task<CustomerDto?> UpdateCustomerAsync(int id, UpdateCustomerDto dto);
-        Task<bool> DeleteCustomerAsync(int id);
+        Task<CustomerDto?> UpdateCustomerAsync(decimal id, UpdateCustomerDto dto);
+        Task<bool> DeleteCustomerAsync(decimal id);
         Task<IEnumerable<CustomerDto>> SearchCustomersAsync(string searchTerm);
     }
 }
