+++ PowerTraderPOS.API/Services/Interfaces/ISupplierService.cs
@@ -5,10 +5,10 @@ namespace PowerTraderPOS.API.Services.Interfaces
     public interface ISupplierService
     {
         Task<IEnumerable<SupplierDto>> GetAllSuppliersAsync();
-        Task<SupplierDto?> GetSupplierByIdAsync(int id);
+        Task<SupplierDto?> GetSupplierByIdAsync(decimal id);
         Task<SupplierDto> CreateSupplierAsync(CreateSupplierDto dto);
-        Task<SupplierDto?> UpdateSupplierAsync(int id, UpdateSupplierDto dto);
-        Task<bool> DeleteSupplierAsync(int id);
+        Task<SupplierDto?> UpdateSupplierAsync(decimal id, UpdateSupplierDto dto);
+        Task<bool> DeleteSupplierAsync(decimal id);
         Task<IEnumerable<SupplierDto>> SearchSuppliersAsync(string searchTerm);
     }
 }
